pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "line-bot-sdk>=3.16.3",
]

[tool.black]
line-length = 100
target-version = ['py310']